# openapi-nodegen-api-file.yml
swagger: '2.0'
info:
  version: 1.0.1
  title: api-dummy_d
  description: A sample API
  contact:
    name: Swagger API Team
    email: john@boats.io
    url: 'https://github.com/johndcarmichael/boats/'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
schemes:
  - http
  - https
host: api.example.com
securityDefinitions:
  apiKey:
    type: apiKey
    in: header
    name: x-api-key
paths:
  /tide:
    get:
      tags:
        - weather
      summary: weather search
      description: Search for weather objects
      operationId: tideGet
      parameters:
        - $ref: '#/parameters/QueryOffset'
        - $ref: '#/parameters/QueryTextSearch'
      responses:
        '200':
          description: Successful fetch
          schema:
            $ref: '#/definitions/WeatherModels'
        '404':
          description: Not found
  /weather:
    get:
      tags:
        - weather
      summary: weather search
      description: Search for weather objects
      operationId: weatherGet
      parameters:
        - $ref: '#/parameters/QueryOffset'
        - $ref: '#/parameters/QueryTextSearch'
      responses:
        '200':
          description: Successful fetch
          schema:
            $ref: '#/definitions/WeatherModels'
        '404':
          description: Not found
    post:
      tags:
        - weather
      summary: weather data
      description: Create a new weather record.
      operationId: weatherPost
      parameters:
        - in: body
          name: weatherPost
          description: Optional description in *Markdown*
          required: true
          schema:
            $ref: '#/definitions/WeatherPost'
      responses:
        '200':
          description: Successful temp creation
          schema:
            $ref: '#/definitions/WeatherModel'
        '404':
          description: Not found
        '422':
          description: Invalid form data provided
      x-permission: apiDummyDCreateWeather
  /weather/latest:
    get:
      tags:
        - weather
      summary: lastest weather data
      description: Get the latest temperatures
      operationId: weatherLatestGet
      produces:
        - application/json
      responses:
        '200':
          description: Successful fetch
          schema:
            $ref: '#/definitions/WeatherModels'
        '404':
          description: Temp not found
  '/weather/{id}':
    delete:
      tags:
        - weather
      summary: weather set to rain
      description: Reset awful sunny weather to excellent rainy weather
      operationId: weatherIdDelete
      x-filename: delete
      parameters:
        - $ref: '#/parameters/PathId'
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found
    get:
      tags:
        - weather
      summary: One weather object
      description: Get the full weather object
      operationId: weatherIdGet
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/PathId'
      responses:
        '200':
          description: Successful fetch
          schema:
            $ref: '#/definitions/WeatherModel'
        '404':
          description: Not found
    put:
      tags:
        - weather
      summary: weather data
      description: Create a new weather record.
      operationId: weatherIdPut
      produces:
        - application/json
      parameters:
        - $ref: '#/parameters/PathId'
        - in: body
          name: weatherIdPut
          description: Optional description in *Markdown*
          required: true
          schema:
            $ref: '#/definitions/WeatherPut'
      responses:
        '200':
          description: Successful temp creation
          schema:
            $ref: '#/definitions/WeatherModel'
        '404':
          description: Not found
        '422':
          description: Invalid form data provided
parameters:
  PathId:
    in: path
    name: id
    type: integer
    required: true
    description: Numeric ID of object to fetch
  QueryOffset:
    in: query
    name: offset
    required: false
    type: integer
    description: The number of items to skip before starting to collect the result set.
  QueryTextSearch:
    in: query
    name: textSearch
    required: false
    type: string
    description: Search string to query
definitions:
  GenericSearchMeta:
    type: object
    properties:
      totalResultCount:
        type: number
      offset:
        type: number
      limit:
        type: number
  JwtAccessModel:
    type: object
    properties:
      userId:
        type: string
  LocationPatch:
    type: object
    properties:
      name:
        type: string
      coordinates:
        type: array
        items:
          type: string
  LocationPost:
    type: object
    required:
      - name
      - coordinates
    properties:
      name:
        type: string
      coordinates:
        type: array
        items:
          type: string
  LocationPut:
    type: object
    required:
      - name
    properties:
      name:
        type: string
  WeatherModel:
    type: object
    properties:
      id:
        type: integer
      date:
        type: string
        format: date
      location:
        type: string
      cloudCoverPercentage:
        type: integer
      humidityPercentage:
        type: integer
      temperature:
        type: number
  WeatherModels:
    type: object
    properties:
      meta:
        $ref: '#/definitions/GenericSearchMeta'
      data:
        type: array
        items:
          $ref: '#/definitions/WeatherModel'
  WeatherPost:
    type: object
    properties:
      date:
        type: string
        format: date
      location:
        type: string
      cloudCoverPercentage:
        type: integer
      humidityPercentage:
        type: integer
      temperature:
        type: number
  WeatherPut:
    allOf:
      - $ref: '#/definitions/WeatherPost'
      - type: object
        properties:
          id:
            type: integer